fix(coding-graph): flip type-source direction so core compiles without the optional peer (#1551 PR1 CI fix)
CI quality job failed because packages/remnic-core/src/coding/optional-coding-graph.ts(33,37) referenced `@remnic/coding-graph` for type-only purposes. In CI's fresh base install (pnpm install --frozen-lockfile), the optional peer package is not symlinked into node_modules, so the DTS build of @remnic/core raised TS2307.

The dependency direction was wrong: @remnic/coding-graph peers on @remnic/core (installed by core), but my PR1 also made core peer on coding-graph + import its types — that creates a circular type-source where core can only compile when the optional package is present.

Fix: types live in core, the optional package imports and implements against them. With this, core's tsup DTS phase emits declarations without resolving `@remnic/coding-graph` at compile time, exactly matching what the base install needs.

Changes:
- packages/remnic-core/src/coding/coding-graph-types.ts (new) — local types-only module: CODING_GRAPH_ENGINE_VERSION, TIER_1_LANGUAGES, CodingGraphLanguage, CodingGraphErrorCode, CodingGraphEngine, ParseFileInput, ParseResult, FileIR, SymbolIR, ImportIR, ExportIR, CallSiteIR, RouteIR, CreateCodingGraphEngineOptions. Zero imports from @remnic/coding-graph.
- packages/remnic-core/src/coding/optional-coding-graph.ts — import the contract types from ./coding-graph-types.js (local) instead of `@remnic/coding-graph`. The dynamic-import loader still uses the computed-specifier pattern; it narrows the `await import(SPECIFIER)` result via `unknown` + a structural isLoadedCodingGraphModule predicate, never via an inline cast on specific fields.
- packages/remnic-core/src/index.ts — re-export CODING_GRAPH_ENGINE_VERSION, TIER_1_LANGUAGES, and the contract types from the main public surface so @remnic/coding-graph (and any future caller) can `import type { ... } from "@remnic/core"`.
- packages/coding-graph/src/index.ts — replace local declarations of TIER_1_LANGUAGES / ENGINE_VERSION / CodingGraphErrorCode type with imports from @remnic/core (single source of truth). CodingGraphEngine, FileIR, SymbolIR, etc. are also imported from @remnic/core; the optional package implements against them and re-exports them under its own name. CodingGraphError remains defined locally because it is runtime (a class, not a type).

Why this preserves the à-la-carte invariant (CLAUDE.md rule 57):
- @remnic/coding-graph still declares @remnic/core in peerDependencies + devDependencies: "workspace:*" (the established connector pattern) so the workspace link is present in dev and the npm metadata is present in publish.
- The dynamic `await import("@remnic/" + "coding-graph")` in core is unchanged: the SPECIFIER is concatenated at runtime so the bundler never bundles the optional package into core, and the loader still throws the user-facing install hint when the optional package is absent.

Verification (matched the CI scenario):
- `pnpm install --no-frozen-lockfile` then check `ls -la node_modules/@remnic/` → only `bench` and `core` symlinks; no coding-graph (matches CI base install).
- `pnpm --filter @remnic/core build` → success, dist/coding/optional-coding-graph.{js,d.ts} emitted, dist/coding-graph-types-*.d.ts emitted. No TS2307.
- `pnpm --filter @remnic/coding-graph build` → success (deps on @remnic/core types via the workspace dev link).

Other gates:
- `npm run preflight:quick` → `[preflight] OK (quick)`.
- `node scripts/check-ratchets.mjs` → `[ratchet] OK`.
- `node --test tests/publish-order.test.mjs` → 5/5 pass (incl. the optional-peer carve-out added in the previous commit).
- `NODE_OPTIONS=--conditions=remnic-source npx tsx --test packages/remnic-core/src/coding/optional-coding-graph.test.ts` → 5/5 pass.
- Base-install integrity: @remnic/coding-graph is absent from any package's `dependencies` and from any tsup noExternal.

PR title unchanged from #1588: feat(coding-graph): scaffold @remnic/coding-graph package + core optional loader (#1551 PR1). Part of #1548 / #1551. Standards #1520.

Chokepoints used: à-la-carte loader pattern (rules 44/57), types-only contract (rule 31), tagged error code (CodingGraphError.code).
