+++ packages/coding-graph/src/index.ts
@@ -1,81 +1,65 @@
 /**
  * @remnic/coding-graph — symbol-extraction engine for codebase memory.
  *
- * À-la-carte optional companion of @remnic/core (CLAUDE.md rule 57). This
- * file ships only the public surface for PR1 (#1551 step 1): the engine
- * placeholder, the version constant, and the not-implemented factory that
- * throws a tagged error so callers (and the core optional loader) always
- * see a *labeled* not-implemented signal rather than a silent stub.
+ * À-la-carte optional companion of @remnic/core (CLAUDE.md rule 57).
  *
- * The real backend (web-tree-sitter parser, per-language extractors,
- * grammar manager) lands in PR2. Until then, `createCodingGraphEngine`
- * returns a tagged `{ ok: false, code: "not_implemented", engineVersion }`
- * result wrapped in a thrown CodingGraphError so call-sites can:
+ * PR1 (#1551, step 1): the package and its build wiring exist. The engine
+ * public surface is declared and the placeholder factory throws a tagged
+ * `CodingGraphError("not_implemented", …)`. The real backend (web-tree-
+ * sitter, per-language extractors, grammar manager) lands in PR2.
  *
- *   (a) detect the placeholder definitively via a structural code ("not_implemented"),
- *   (b) surface the engine VERSION the user requested (helpful once PR2 ships),
- *   (c) propagate it as a normal thrown error for try/catch flow.
+ * Type-source direction:
+ *   The contract types (CodingGraphEngine, FileIR, etc.) and the
+ *   `TIER_1_LANGUAGES` / `ENGINE_VERSION` constants live in
+ *   @remnic/core (packages/remnic-core/src/coding/coding-graph-types.ts
+ *   and `CODING_GRAPH_ENGINE_VERSION` re-export from the main index).
+ *   This package imports them and implements against them; it does NOT
+ *   redefine them. That keeps a single source of truth so core's tsup
+ *   DTS phase can emit declarations in CI's fresh base install (where
+ *   this optional peer is not symlinked at compile time).
  *
- * The grammar .wasm assets directory (`grammars/`) is listed in
- * package.json `files` so PR2's `.wasm` payloads ship in the tarball
- * without further packaging changes.
+ *   @remnic/coding-graph declares @remnic/core as both `peerDependencies`
+ *   and `devDependencies: "workspace:*"` in its package.json, so the
+ *   pnpm workspace link exists and the `import from "@remnic/core"`
+ *   below resolves in development.
  */
 
+import {
+  TIER_1_LANGUAGES,
+  type CodingGraphEngine,
+  type CodingGraphErrorCode,
+  type CodingGraphLanguage,
+  type CreateCodingGraphEngineOptions,
+  type FileIR,
+  type ParseFileInput,
+  type ParseResult,
+  type SymbolIR,
+} from "@remnic/core";
+
 // ---------------------------------------------------------------------------
-// Version constant
+// Engine version constant — kept under the `ENGINE_VERSION` name (and
+// ALSO exposed by @remnic/core as `CODING_GRAPH_ENGINE_VERSION`) so the
+// dynamic-import loader in @remnic/core can validate the shape via a
+// single field name without an alias dance. The constant must match
+// across both packages or the loader's structural check fails.
 // ---------------------------------------------------------------------------
 
-/**
- * Public engine version. Bumped in lockstep with the package version while
- * the engine is under active development (PR2+). External callers should
- * treat this as informational; semantic compatibility is not promised
- * until v10 ships.
- */
+/** Public engine version (single source of truth: @remnic/core). */
 export const ENGINE_VERSION = "0.1.0-pr1" as const;
 
 // ---------------------------------------------------------------------------
-// Tier 1 language list (declared now; tree-sitter grammar .wasm files
-// themselves ship in PR2)
+// Tier-1 language list re-export. The list itself lives in @remnic/core
+// (single source of truth) so consumers that reach the optional package
+// via the loader get the same shape they would get from core directly.
 // ---------------------------------------------------------------------------
-
-/**
- * Tier-1 languages the engine will support when implementation lands.
- * Order is stable; per-language configuration is not yet exposed because
- * no extractor exists yet — adding config in PR1 would force PR2 to
- * thread it through every IR site.
- */
-export const TIER_1_LANGUAGES = [
-  "typescript",
-  "tsx",
-  "javascript",
-  "python",
-  "go",
-  "rust",
-  "java",
-  "c",
-  "cpp",
-  "csharp",
-  "ruby",
-  "php",
-  "kotlin",
-  "swift",
-  "bash",
-] as const;
-
-export type CodingGraphLanguage = (typeof TIER_1_LANGUAGES)[number];
+export { TIER_1_LANGUAGES };
+export type { CodingGraphLanguage };
 
 // ---------------------------------------------------------------------------
-// Tagged error — carriers of the not-implemented signal
+// Tagged error
 // ---------------------------------------------------------------------------
 
-/**
- * Stable error code so callers (and the optional loader in core) can
- * pattern-match without parsing human-readable strings. New codes must
- * be added here so consumers see them via TypeScript.
- */
-export type CodingGraphErrorCode =
-  | "not_implemented"
-  | "module_load_failed";
+export type { CodingGraphErrorCode } from "@remnic/core";
 
 /**
  * Thrown by `createCodingGraphEngine` while the real implementation is
@@ -99,102 +83,7 @@ export class CodingGraphError extends Error {
 }
 
 // ---------------------------------------------------------------------------
-// Engine interface — final shape PR2 implements against
-// ---------------------------------------------------------------------------
-
-/**
- * Public engine surface. The interface itself is complete enough for PR2
- * to satisfy without churn; the placeholder implementation throws on
- * construction, so the runtime contract is observable today.
- */
-export interface CodingGraphEngine {
-  /** Engine version reported at construction time. */
-  readonly engineVersion: string;
-  /** Tier-1 languages this build supports (PR2 will narrow by grammar availability). */
-  readonly supportedLanguages: readonly CodingGraphLanguage[];
-  /**
-   * Parse a single source file and emit its FileIR.
-   *
-   * PR1 always throws `CodingGraphError("not_implemented", ...)`. PR2 will
-   * return `ParseResult`; failure paths come back as
-   * `{ ok: false, code: "parse_failed", path, message }` (rule 44) rather
-   * than partial / silent IR.
-   */
-  parseFile(input: ParseFileInput): Promise<ParseResult>;
-  /** Engine lifecycle — release any cached parsers/grammars. */
-  dispose(): Promise<void>;
-}
-
-export interface ParseFileInput {
-  /** Repository-relative path (forward slashes; no leading `./`). */
-  readonly path: string;
-  /** Raw file bytes; hashing happens inside the engine (rule 23). */
-  readonly content: Uint8Array;
-  /**
-   * Optional override. When omitted the engine sniffs the language from
-   * `path` extensions against its built-in tier-1 list (PR2).
-   */
-  readonly language?: CodingGraphLanguage;
-}
-
-export type ParseResult =
-  | { readonly ok: true; readonly ir: FileIR }
-  | { readonly ok: false; readonly code: "parse_failed"; readonly path: string; readonly message: string };
-
-/**
- * Intermediate representation contract — the seam between the parser
- * (this package) and the graph store (sibling issue #1552). PR1 declares
- * the shape; PR2 fills the extractor that produces one. Field types are
- * the minimal viable set per the issue's design section.
- */
-export interface FileIR {
-  readonly path: string;
-  readonly language: CodingGraphLanguage;
-  /** SHA-256 of the raw bytes; rule 23 — every consumer hashes the same form. */
-  readonly contentHash: string;
-  readonly symbols: readonly SymbolIR[];
-  readonly imports: readonly ImportIR[];
-  readonly exports: readonly ExportIR[];
-  readonly callSites: readonly CallSiteIR[];
-  readonly routes: readonly RouteIR[];
-}
-
-export interface SymbolIR {
-  readonly kind: "function" | "class" | "method" | "interface" | "enum" | "type" | "module";
-  readonly name: string;
-  readonly qualifiedName: string;
-  /** Half-open byte span `[startByte, endByte)`. Rule 35. */
-  readonly span: { readonly startByte: number; readonly endByte: number };
-  readonly parentQualifiedName?: string;
-}
-
-export interface ImportIR {
-  /** Raw module specifier as written in source. */
-  readonly module: string;
-  readonly importedNames: readonly string[];
-  readonly span: { readonly startByte: number; readonly endByte: number };
-}
-
-export interface ExportIR {
-  readonly name: string;
-  readonly span: { readonly startByte: number; readonly endByte: number };
-}
-
-export interface CallSiteIR {
-  readonly calleeNameCandidates: readonly string[];
-  readonly span: { readonly startByte: number; readonly endByte: number };
-}
-
-export interface RouteIR {
-  /** HTTP verb in upper-case, or framework-native verb (e.g. "ANY"). */
-  readonly verb: string;
-  readonly pathTemplate: string;
-  readonly handlerQualifiedName: string;
-  readonly span: { readonly startByte: number; readonly endByte: number };
-}
-
-// ---------------------------------------------------------------------------
-// Placeholder factory
+// Engine implementation for PR1 (placeholder — PR2 fills extractors).
 // ---------------------------------------------------------------------------
 
 /**
@@ -218,11 +107,17 @@ export function createCodingGraphEngine(
   );
 }
 
-/**
- * Reserved for PR2. Declared now so the public surface is stable; the
- * options object is intentionally empty in PR1.
- */
-export interface CreateCodingGraphEngineOptions {
-  /** Reserved for PR2: extra grammar directory supplied by the operator. */
-  readonly grammarDir?: never;
-}
+// ---------------------------------------------------------------------------
+// Re-export the contract types so the public surface is stable whether a
+// consumer reaches into @remnic/coding-graph or @remnic/core. `export type`
+// guarantees these are erased at runtime — no double-emit.
+// ---------------------------------------------------------------------------
+
+export type {
+  CodingGraphEngine,
+  CreateCodingGraphEngineOptions,
+  FileIR,
+  ParseFileInput,
+  ParseResult,
+  SymbolIR,
+};
